Handle updating period inside of modal
